% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gseaLine.R
\name{geom_gseaLine}
\alias{geom_gseaLine}
\title{GSEA line for ggplot2}
\usage{
geom_gseaLine(df, ...)
}
\arguments{
\item{df}{data.frame, calculated by the function gseaCurve}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a ggplot layer that can be added to a ggplot object
}
\description{
Imports:
ggplot2
}
\details{
uses the ggplot2::layer function, with geom="path"
}
\examples{
library(gggsea)
library(ggplot)

curve <- gseaCurve()

}
