% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gsea.R
\name{theme_gsea}
\alias{theme_gsea}
\title{applies a GSEA-friendly theme to a ggplot object}
\usage{
theme_gsea(textsize = 14)
}
\arguments{
\item{textsize}{number, sets the size for every text in the plot, except those that are defined extra (e.g. annotate)}
}
\value{
a ggplot layer that can be added to a ggplot object
}
\description{
Imports:
ggplot2
}
\details{
sets options for the panel.grid, panel.background, panel.border, strip.background, strip.text, text, axis.text, axis.title and plot.title
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_hm()

}
