% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gsea.R
\name{geom_gsea}
\alias{geom_gsea}
\title{create a venn diagram in ggplot2}
\usage{
geom_gsea(df, withTicks = T, withGradient = T, prettyGSEA = T, ...)
}
\arguments{
\item{df}{data.frame, calculated by the function gseaCurve}

\item{withTicks}{boolean, should there be ticks?}

\item{withGradient}{boolean, should ther be a color gradient?}

\item{prettyGSEA}{boolean, should some aesthetics be automatically added? Adds a 0-line, regulates y-breaks and labels the axes.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a list of ggplot layers, ready to be added to a ggplot object
}
\description{
Imports:
ggplot2
}
\details{
uses the functions geom_gseaLine, geom_gseaTicks and geom_gseaGradient, from this package
}
\examples{
library(gggsea)
library(ggplot)

curve <- gseaCurve()

}
